Add stage task, make app portable by fixing URL
The API URL should not be hardcoded to localhost, instead it should
point to a relative path. It will allow to find the API on the same
machine regardless of the way the app is run.

Gradle 'stage' task is needed for Heroku deployment.

Add Procfile to define how the production app should run.