Fix date timezone
When selecting some date from a datepicker it is selected based on timezone
local to the browser. When building a request this local date is converted to
UTC. Previously the backend API was storing only Date information, so on receiving
data from client timezone info was stripped. Therefore a date sent back in response
was often incorrect, as it was regular date without time and timezone mapped back to
date with time in UTC.

Right now the backend API accepts DateTime with timezone, and stores Date in UTC.
When any DateTime data is send back to the client, it is always in UTC, therefore it is
client's responsibility to handle locale in a correct way.